requirements.txt
streamlit>=1.32
pymupdf>=1.23
sentence-transformers>=2.6
faiss-cpu>=1.8.0
numpy>=1.24
requests>=2.31
